backend/requirements.txt
langchain~=0.3.27
langchain-core
langchain-community
langchain-text-splitters
pypdf
docx2txt
fastapi~=0.116.1
uvicorn
streamlit
python-dotenv~=1.1.1
pinecone~=7.3.0
langchain-pinecone
langchain-groq
requests
uuid
langchain-huggingface
pydantic~=2.11.7
pillow
google-cloud-vision
google-auth
voyageai
pinecone-client
torch~=2.8.0
dnspython~=2.7.0
email-validator~=2.3.0
PyYAML~=6.0.2
tqdm~=4.67.1
pdf2image~=1.17.0
numpy~=2.2.6

